+++ tests/BoltOn.Tests/Data/CosmosDb/TestSchoolCosmosDbOptions.cs
@@ -1,0 +1,8 @@
+﻿using BoltOn.Data.CosmosDb;
+
+namespace BoltOn.Tests.Data.CosmosDb
+{
+    public class TestSchoolCosmosDbOptions : BaseCosmosDbOptions
+    {
+    }
+}
